pyproject.toml
[project]
name = "maxlog"
version = "0.0.1"
description = "Loguru logging for personal projects"
authors = [
    {name = "Max Ludden", email = "dev@maxludden.com"},
]
dependencies = [
    "loguru>=0.7.2",
    "ruff>=0.1.7",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
